refactor(calm-hub): address Will Osborne PR review comments on MappingControllerResource
- Extract MappingControllerService (@ApplicationScoped) containing all
  business logic: store interactions, createOrAddVersion, updateVersioned
  Resource, getVersionsForMapping, getResourceJsonForVersion, control
  requirement/configuration handling, resolveControlId/resolveConfigId
- MappingControllerResource reduced from ~1250 to ~430 lines; now a thin
  HTTP layer responsible only for JAX-RS wiring, permission checks, and
  exception-to-Response mapping
- Replace all inline @jakarta.validation.constraints.Pattern FQNs with
  an import and shorthand @Pattern (covers ~12 sites)
- Remove duplicated getMapping->createOrAddVersion block from both
  createResourceFromDocument and handlePost; both now delegate to
  service.createOrAddVersion()

Verified: 1945/1945 unit tests pass, 490/490 integration tests pass
